requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Anthropic API
anthropic>=0.18.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
PyMuPDF>=1.23.0

# Web interface
streamlit>=1.30.0

# Environment variables
python-dotenv>=1.0.0

# Additional dependencies that may be needed
numpy>=1.24.0
pandas>=2.0.0
typing-extensions>=4.5.0